refactor: Split app_state.rs into separate modules
app_state.rs was 768 lines, now split into:
- service/app_state.rs (~65 lines): AppState, ActiveViewer, ActiveTab, common utils
- service/viewer_state.rs (~390 lines): ViewerState, FileTree, tree operations
- service/explorer_state.rs (~365 lines): ExplorerState, Thumbnail, pagination

Updated imports in:
- app/mod.rs
- app/explorer.rs
- app/explorer_helpers.rs
- app/viewer.rs
- utils/watcher_utils.rs
