+++ src-tauri/src/service/mod.rs
@@ -1,1 +1,3 @@
 pub mod app_state;
+pub mod explorer_state;
+pub mod viewer_state;
